fix: rename Unstruct data labeler profile to unstruct labeler profile
